Servicable abstraction of network rule display into separate file
Need to clean up hackery in MatchedNetworkProtocolDataTemplate; I can't keep adding classes depending on the context I call it in.  Maybe it needs to be a control as well
Need to fix display of headers
Try switching a few other views over to using it?  Need to find a way to have it coexist with the Overlap summaries, since I need the OverlapType to also show on the right