Checkpoint of wildly ugly, partially working solution to drift issues between rules
Lots of drift was occuring when large numbers of very small ranges were next to each other, each triggering the MinDisplaySize rule
This would cause them to add up to be larger than the equivalently-sized consolidated rule
That meant that rules being compared against those ranges would be shifted left, because they consolidated the comparison ranges and therefore produced a smaller width
Trying to resolve that by calculating the size of the ranges from the source rule and using that to draw ranges.  That works when the source rule has the large numbers of small rules, but not when the comparison rule does.  Also throws off the overall width calc because the overall width calc also functions by consolidated ranges.  Gotta redesign this, frankly