Add epicycles to stop right-aligning ranges inside larger containers
The internalRatio calculation assumed that the ending of the largest range should be placed at the end of the container, effectively right-aligning it.
Added some logic to only apply that when the contained range size was larger than the container (which only happens when MinDisplaySize is in play).
Added logic to track a separate DistanceTravelled metric for each level of Depth so that cases with a range at a different depth than a collection of single IPs doesn't cause the range to be squashed just because of the single IPs